requirements.txt
faster-whisper
ffmpeg-python
fire
fschat
openai-whisper
python-dotenv
python-youtube
rich
setuptools-rust
tqdm